Flight/Guidance: Initial template for guidance module Flight: Create PositionDesired (the active waypoint) UAVObject and make the FlightSituationActual no update since it not used. Flight: New velocity desired object that passes information between the look computing the desired velocity and the PID loop to get it (updated at different rates) UAVObjects/PositionActual: Remove unused GPS fields UAVObjects/PositionActual VelocityActual: Split the velocity into a separate object.  ALso make sure all the information telemetered around is in cm to avoid using floats. UAVObject/GuidanceSettings: New guidance settings object for the guidance module Flight/Posix: Add the new objects to the Posix sim Flight/Guidance: Computes a desired velocity based on position error than runs a PID loop to control roll and pitch to achieve that velocity.  All distances are in cm, and updated the PositionActual fields to reflect this and use int32.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@1760 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ flight/PiOS/inc/pios_opahrs_proto.h
@@ -319,8 +319,8 @@ struct opahrs_msg_v1_rsp_update {
     float q3;
     float q4;
   } quaternion;
-  float NED[3];
-  float Vel[3];
+  int32_t NED[3];
+  int32_t Vel[3];
   uint8_t load;
   uint8_t idle_time;
   uint8_t run_time;
